fix mcp scenario examples
